refactor(go-react-spa): server lifecycle を internal/server.Run(ctx) に分離
`cmd/server/main.go` に直書きだった logger / Config / DB / HTTP server +
graceful shutdown 一式を `internal/server` パッケージの
`Run(ctx context.Context) error` に切り出す。`main()` は
`signal.NotifyContext` でシグナル付き ctx を作り `server.Run()` を呼ぶ
だけの薄ラッパに簡素化。

これにより既存 cobra CLI に「サーバ起動」サブコマンドとして組み込む際:

    RunE: func(cmd *cobra.Command, _ []string) error {
        return server.Run(cmd.Context())
    }

の形で 1 行 import するだけで済む。scaffold 経路 (`./bin/server`) の
起動シーケンスとシグナル経由の graceful shutdown 挙動は現行同一
（実バイナリで SIGTERM → exit 0 を確認済み）。

設計判断:
- `Run()` シグネチャは `(ctx) error` で固定。`*Config` を引数で受ける
  形は将来の拡張ポイントとして残す
- ListenAndServe goroutine の error は buffered channel + select で
  伝播。`os.Exit(1)` を内部から呼ぶ経路は撤廃し、cobra ユーザが自分の
  エラー処理を貫通させられるようにした
- `internal/server` 配下に置いたので同一 module 内の cobra からのみ
  import 可能。公開 API 化はまだコミットせず、`--pick` で取り込む
  ワークフローを前提にする

DoD 充足:
- `make build && ./bin/server` の起動 / graceful shutdown は現行同一
- `internal/server/server_test.go` の `TestRun_ShutsDownOnCtxCancel`
  が `t.Setenv("PORT","0")` / `DATABASE_DSN=":memory:"` で PASS
- README に「cobra subcommand への組み込み」セクションを追加
- `download.sh go-react-spa-noauth` で scaffold した結果でも
  `Run()` 構造が維持され、`go test ./...` が PASS することを確認

Refs: #219
